Review: don't default to unit name if source-name isn't provided
We were previously defaulting to the unit name if a Grafana source was provided, but since unit names can frequently change, we will default to the <app_name>_<relation_id>. This way, we can also safely assume a unique name is always available.

This commit also adjusts tests cases to be in line with the new logic.