Be more resilient to timing with the unit address (#54)
It's possible for network.get_binding to return `None` if an
address is not set, which results in a source address which is
something like `None:9090` when interpolated.

If the address can't be retrieved, don't set it in the unit data
bucket. The refresh event (by default, `pebble_ready`) will catch
this. This helps us avoid confusing/breaking the Grafana database
by submitting multiple provisioning events which can include
adding then immediately removing a source from `deleteDatasources`
if a relation is rapidly broken/re-joined.

While we're at it, hash the values by pulling from the container
and rework the charm tests since the Pebble harness works