Do a lot more work on dashboard templates
Return an appropriate error to the Provider if the JSON is bad.

If there are no template variables, assume any datasources are
prometheus (impossible to tell from `panel` objects), and replace
their values with `${prometheusds}`

If there **are** template variables, sift through them. If they
are known types for Loki or Prometheus, replace them by name in the
templates, then go back and clean them up from template variables.

Add tests for all of this, obviously.