Update the Grafana dashboard docs, accept _any_ file
Since we're not really using Jinja templates anymore, accept any
file contained in the template path, and log warning messages if
there's invalid JSON (validating even a small subset of what could
be a "valid" dashboard is practically an NP-complete problem, so
we can guess that if it's valid JSON, it's a valid dashboard).

Update the unit tests to match the change.

Make a one-line change to the docs to add `requires:` for the
metadata in providers.

Closes #77
Closes #78