Add sqlite streaming from primaries to secondaries
Considering that is it not possible to `limit` the number of units
which may be started in a Juju application, it's possible for
more than one Grafana unit to be initialized. MOST of the data
which is present is propagated though libraries and across relation
interfaces as dashboards/datasources, but user interaction and
logins with the Grafana UI itself may result in changes.

Without a shared database, scaling the application up or down could
result in units with different administrative passwords, dashboards
imported from the marketplace, manually added users/groups, or more.

When a charm is elected as the leader, it will now start `litestream`
to serve changes over GRPC/HTTP. This additionally requires changing
the a sqlite pragma to Grafana. However, this is a long-requested
feature for Grafana itself, with hundreds of users reporting that
it actually reduced the frequency of database locking [see the
issue](https://github.com/grafana/grafana/issues/16638), so this is
reasonably safe.