Allow configurable datasource query timeouts
This commit adds a configuration parameter to the Grafana charm
that allows setting a minimum default timeout for querrying
data sources. This is necessary because some data sources such
as Loki can take long time to return query responses. If the
Grafana query timeout (which defaults to 30s) is not set suitably
this may lead to 5xx HTTP errors.