Datasource Query timeout Configuration (#107)
* Allow configurable datasource query timeouts

This commit adds a configuration parameter to the Grafana charm
that allows setting a minimum default timeout for querrying
data sources. This is necessary because some data sources such
as Loki can take long time to return query responses. If the
Grafana query timeout (which defaults to 30s) is not set suitably
this may lead to 5xx HTTP errors.

* Added unit test datasource timeout config

This commit adds unit tests for the datasource query timeout
configuration. The unit tests check
- that configuration value overrides timeout value set in relation
  data if the configuration value is larger.
- that configuration value is overridden by relation data timeout
  value if the latter is larger.