Add ingress support
Neither ingress-per-unit nor ingress-per-app do precisely what
Grafana needs in order to keep sqlite in a reasonable state, but
the TraefikRoute libraries, as long as we ignore the work done by
the actual `traefik-route-k8s` charm entirely, provide an interface
via which we can send "raw" Traefik config snippets to ingress.

`web-external-url` was already an available config option used by
the load tests, so it is reasonably exercised and tested already,
and matches the existing expected semantics for ingress on other
charms.

If a relation is established to Traefik without `web-external-url`
being set, the Grafana leader will configure Traefik with a route
which points to `{model}-{app.name}` publicly, which will use the
`{app}-{unit_num}-{endpoints}` binding Juju gives us in order to
forward external traffic to the leader.

If `web-external-url` is set, use that, with the same internal
endpoints as above.

On leadership election, rewrite the route so it points at the
correct leader. Similarly, if a config change sets `web-external-url`,
use that.

Since `TraefikRouteRequirer` has a non-optional `Relation` as part
of the constructor, _also_ configure ingress if a relation is
established during the middle of a charm lifecycle after setting
a private variable in the library.