Fix type checking issue with service_ports
This worked in Pyright 1.1.314, but not in 1.1.315 or 1.1.316.

The type of service_ports is
List[Tuple[str, Literal[3000], Literal[3000]]], which doesn't type
check as List[Tuple[str, int, int]] -- which is very unintuitive, but
oh well.

I believe this is by design, see
https://github.com/microsoft/pyright/issues/5366 and
https://microsoft.github.io/pyright/#/type-inference?id=tuple-expressions

The Pyright fixes were probably one of
https://github.com/microsoft/pyright/issues/5321 or
https://github.com/microsoft/pyright/pull/5323