refactor: move the grafana UID defintion to the charm from grafana_source
The grafana_source relation sends a unique string identifying this grafana.  Previously, this string was created in the grafana_source library, making it a bit awkward to use in other libraries.

This PR proposes a change where this identifier is created/owned by the grafana charm rather than the datasource lib.  Outwardly, this should cause no change in function.