refactor: move the grafana UID definition to the charm from grafana_source (#387)
* refactor: move the grafana UID defintion to the charm from grafana_source

The grafana_source relation sends a unique string identifying this grafana. Previously, this string was created in the grafana_source library, making it a bit awkward to use in other libraries.

This PR proposes a change where this identifier is created/owned by the grafana charm rather than the datasource lib. Outwardly, this should cause no change in function. The motivation for this change is to use this same identifier in the forthcoming grafana-metadata relation

The grafana_uid is meant for identifying an application of charmed grafana, so it should be universally unique (unique even across different models, deployments, etc). Given that, it really is owned by grafana the application, and the related lib merely uses that ID.

This change introduces a breaking change in the library API due to an additional required parameter. The libapi has not been bumped here because grafana-k8s is the only operator that consumes this (and because, if there is someone else consuming it, the change in api will result in an obvious bug rather than a hidden bug, so the risk of real damage is low).